+++ src/main/java/com/senla/project/socialnetwork/service/BlocklistService.java
@@ -1,36 +1,35 @@
 package com.senla.project.socialnetwork.service;
 
 import com.senla.project.socialnetwork.entity.Blocklist;
-import com.senla.project.socialnetwork.exeptions.NoSuchIdException;
+import com.senla.project.socialnetwork.exeptions.NoAccountsException;
+import com.senla.project.socialnetwork.exeptions.NoSuchElementException;
 import com.senla.project.socialnetwork.repository.BlocklistRepository;
-import lombok.SneakyThrows;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class BlocklistService {
 
     private BlocklistRepository blocklistRepository;
 
-    public BlocklistService(BlocklistRepository blocklistRepository) {
-        this.blocklistRepository = blocklistRepository;
-    }
-
-    public void add(Blocklist blocklist){
+    public void add(Blocklist blocklist) {
         blocklistRepository.save(blocklist);
     }
 
-    public List<Blocklist> findAll(){
+    public List<Blocklist> findAll() {
+        if (blocklistRepository.findAll().isEmpty()) {
+            throw new NoAccountsException();
+        }
         return blocklistRepository.findAll();
     }
 
-    @SneakyThrows
-    public Blocklist findById(Long id){
-        return blocklistRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public Blocklist findById(Long id) {
+        return blocklistRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
-    @SneakyThrows
     public Blocklist update(Long id, Blocklist blocklist) {
 
         return blocklistRepository.findById(id).map(bl -> {
@@ -41,12 +40,13 @@ public class BlocklistService {
             bl.setBlockCause(blocklist.getBlockCause());
             return blocklistRepository.save(bl);
         })
-                .orElseThrow(() ->
-                        new NoSuchIdException(id, "blocked user")
-                );
+                .orElseThrow(NoSuchElementException::new);
     }
 
     public void delete(Long id) {
+        if (blocklistRepository.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         blocklistRepository.deleteById(id);
     }
 }
