+++ src/main/java/com/senla/project/socialnetwork/service/ContactService.java
@@ -1,36 +1,35 @@
 package com.senla.project.socialnetwork.service;
 
 import com.senla.project.socialnetwork.entity.Contact;
-import com.senla.project.socialnetwork.exeptions.NoSuchIdException;
+import com.senla.project.socialnetwork.exeptions.NoAccountsException;
+import com.senla.project.socialnetwork.exeptions.NoSuchElementException;
 import com.senla.project.socialnetwork.repository.ContactRepository;
-import lombok.SneakyThrows;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class ContactService {
 
     private ContactRepository contactRepository;
 
-    public ContactService(ContactRepository contactRepository) {
-        this.contactRepository = contactRepository;
-    }
-
-    public void add(Contact contact){
+    public void add(Contact contact) {
         contactRepository.save(contact);
     }
 
-    public List<Contact> findAll(){
+    public List<Contact> findAll() {
+        if (contactRepository.findAll().isEmpty()) {
+            throw new NoAccountsException();
+        }
         return contactRepository.findAll();
     }
 
-    @SneakyThrows
-    public Contact findById(Long id){
-        return contactRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public Contact findById(Long id) {
+        return contactRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
-    @SneakyThrows
     public Contact update(Long id, Contact contact) {
 
         return contactRepository.findById(id).map(cont -> {
@@ -41,12 +40,13 @@ public class ContactService {
             cont.setContactRole(contact.getContactRole());
             return contactRepository.save(cont);
         })
-                .orElseThrow(() ->
-                        new NoSuchIdException(id, "contact")
-                );
+                .orElseThrow(NoSuchElementException::new);
     }
 
     public void delete(Long id) {
+        if (contactRepository.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         contactRepository.deleteById(id);
     }
 }
