+++ src/main/java/com/senla/project/socialnetwork/service/MessageService.java
@@ -1,36 +1,35 @@
 package com.senla.project.socialnetwork.service;
 
 import com.senla.project.socialnetwork.entity.Message;
-import com.senla.project.socialnetwork.exeptions.NoSuchIdException;
+import com.senla.project.socialnetwork.exeptions.NoAccountsException;
+import com.senla.project.socialnetwork.exeptions.NoSuchElementException;
 import com.senla.project.socialnetwork.repository.MessageRepository;
-import lombok.SneakyThrows;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class MessageService {
 
     private MessageRepository messageRepository;
 
-    public MessageService(MessageRepository messageRepository) {
-        this.messageRepository = messageRepository;
-    }
-
-    public void add(Message message){
+    public void add(Message message) {
         messageRepository.save(message);
     }
 
-    public List<Message> findAll(){
+    public List<Message> findAll() {
+        if (messageRepository.findAll().isEmpty()) {
+            throw new NoAccountsException();
+        }
         return messageRepository.findAll();
     }
 
-    @SneakyThrows
-    public Message findById(Long id){
-        return messageRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public Message findById(Long id) {
+        return messageRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
-    @SneakyThrows
     public Message update(Long id, Message message) {
 
         return messageRepository.findById(id).map(mess -> {
@@ -40,12 +39,13 @@ public class MessageService {
             mess.setMessageTxt(message.getMessageTxt());
             return messageRepository.save(mess);
         })
-                .orElseThrow(() ->
-                        new NoSuchIdException(id, "message")
-                );
+                .orElseThrow(NoSuchElementException::new);
     }
 
     public void delete(Long id) {
+        if (messageRepository.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         messageRepository.deleteById(id);
     }
 }
