+++ src/main/java/com/senla/project/socialnetwork/service/RoleListService.java
@@ -1,10 +1,10 @@
 package com.senla.project.socialnetwork.service;
 
 import com.senla.project.socialnetwork.entity.RoleList;
-import com.senla.project.socialnetwork.exeptions.NoSuchIdException;
+import com.senla.project.socialnetwork.exeptions.NoAccountsException;
+import com.senla.project.socialnetwork.exeptions.NoSuchElementException;
 import com.senla.project.socialnetwork.repository.RoleListRepository;
 import lombok.AllArgsConstructor;
-import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,32 +15,34 @@ public class RoleListService {
 
     private RoleListRepository roleListRepository;
 
-    public void add(RoleList roleList){
+    public void add(RoleList roleList) {
         roleListRepository.save(roleList);
     }
 
-    public List<RoleList> findAll(){
+    public List<RoleList> findAll() {
+        if (roleListRepository.findAll().isEmpty()) {
+            throw new NoAccountsException();
+        }
         return roleListRepository.findAll();
     }
 
-    @SneakyThrows
-    public RoleList findById(Long id){
-        return roleListRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public RoleList findById(Long id) {
+        return roleListRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
-    @SneakyThrows
     public RoleList update(Long id, RoleList roleList) {
 
         return roleListRepository.findById(id).map(rl -> {
             rl.setName(roleList.getName());
             return roleListRepository.save(rl);
         })
-                .orElseThrow(() ->
-                        new NoSuchIdException(id, "role list")
-                );
+                .orElseThrow(NoSuchElementException::new);
     }
 
     public void delete(Long id) {
+        if (roleListRepository.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         roleListRepository.deleteById(id);
     }
 }
