+++ src/main/java/com/senla/project/socialnetwork/service/UserOfCommunityService.java
@@ -1,36 +1,35 @@
 package com.senla.project.socialnetwork.service;
 
 import com.senla.project.socialnetwork.entity.UserOfCommunity;
-import com.senla.project.socialnetwork.exeptions.NoSuchIdException;
+import com.senla.project.socialnetwork.exeptions.NoAccountsException;
+import com.senla.project.socialnetwork.exeptions.NoSuchElementException;
 import com.senla.project.socialnetwork.repository.UserOfCommunityRepository;
-import lombok.SneakyThrows;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@AllArgsConstructor
 public class UserOfCommunityService {
 
     private UserOfCommunityRepository userOfCommunityRepository;
 
-    public UserOfCommunityService(UserOfCommunityRepository userOfCommunityRepository) {
-        this.userOfCommunityRepository = userOfCommunityRepository;
-    }
-
-    public void add(UserOfCommunity userOfCommunity){
+    public void add(UserOfCommunity userOfCommunity) {
         userOfCommunityRepository.save(userOfCommunity);
     }
 
-    public List<UserOfCommunity> findAll(){
+    public List<UserOfCommunity> findAll() {
+        if (userOfCommunityRepository.findAll().isEmpty()) {
+            throw new NoAccountsException();
+        }
         return userOfCommunityRepository.findAll();
     }
 
-    @SneakyThrows
-    public UserOfCommunity findById(Long id){
-        return userOfCommunityRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public UserOfCommunity findById(Long id) {
+        return userOfCommunityRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
 
-    @SneakyThrows
     public UserOfCommunity update(Long id, UserOfCommunity userOfCommunity) {
 
         return userOfCommunityRepository.findById(id).map(uoc -> {
@@ -39,12 +38,13 @@ public class UserOfCommunityService {
             uoc.setDateEntered(userOfCommunity.getDateEntered());
             return userOfCommunityRepository.save(uoc);
         })
-                .orElseThrow(() ->
-                        new NoSuchIdException(id, "user in community")
-                );
+                .orElseThrow(NoSuchElementException::new);
     }
 
     public void delete(Long id) {
+        if (userOfCommunityRepository.findById(id).isEmpty()) {
+            throw new NoSuchElementException();
+        }
         userOfCommunityRepository.deleteById(id);
     }
 }
